% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFileCache.R
\name{checkLocalFileRepo}
\alias{checkLocalFileRepo}
\title{Check if geneinfo2db_local_filecache direcory exists. If not creates this path.}
\usage{
checkLocalFileRepo(subdir = getOption("geneinfo2db_local_filecache"))
}
\arguments{
\item{subdir}{direcotry for storing local copies of remote files.}
}
\value{
path
}
\description{
Check if geneinfo2db_local_filecache direcory exists. If not creates this path.
}
\examples{

\donttest{
checkLocalFileRepo(tempdir())
}

}
