% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDatabase.R
\name{createDatabase}
\alias{createDatabase}
\title{creates parts of a Postgresql database}
\usage{
createDatabase(db_part)
}
\arguments{
\item{db_part}{a part of the database structure stored in this package}
}
\value{
nothing
}
\description{
creates parts of a Postgresql database
}
