% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeRead.R
\name{download_filesize}
\alias{download_filesize}
\title{Get file size info from the URL.}
\usage{
download_filesize(URL)
}
\arguments{
\item{URL}{URL to test. Can be a ftp URL.}
}
\value{
file size of the file from the URL.
}
\description{
Get file size info from the URL.
}
\examples{


\dontrun{
download_filesize("https://ftp.ncbi.nlm.nih.gov/refseq/H_sapiens/Homo_sapiens.gene_info.gz")
}
}
