% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCelllineAnnotation.R
\name{getCelllineAnnotation}
\alias{getCelllineAnnotation}
\title{getCelllineAnnotation}
\usage{
getCelllineAnnotation()
}
\value{
a list containing two data frames: 
one with cell line annotations and another with alternative cell line names
}
\description{
getCelllineAnnotation
}
