% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSettings.R
\name{getDBConfig}
\alias{getDBConfig}
\title{Get password from .pgpass file}
\usage{
getDBConfig(dbsettings)
}
\arguments{
\item{dbsettings}{incomplete list of dbhost, dbname, dbuser, dbport, or dbpass}
}
\value{
complete list with database configuration.
}
\description{
Get password from .pgpass file
}
\note{
the return value of this function is used to specify DB config for \code{setDBconfig}.
}
