% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFileCache.R
\name{getLocalFileRepo}
\alias{getLocalFileRepo}
\title{Returns the path for local file cache.}
\usage{
getLocalFileRepo()
}
\value{
Path to local file cache.
}
\description{
Returns the path for local file cache.
}
\details{
This function also checks if the local cache exists by calling \code{\link{checkLocalFileRepo}}.
}
\examples{

\dontrun{
getLocalFileRepo()
}

}
