% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigDB.R
\name{getMSigDB}
\alias{getMSigDB}
\title{writes a list of data frames into a Postgresql database}
\usage{
getMSigDB(version)
}
\value{
a list with a single list element called public.msigdb
}
\description{
writes a list of data frames into a Postgresql database
}
