% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostgresqlConnection.R
\name{getPostgresqlConnection}
\alias{getPostgresqlConnection}
\title{Get connection to postresql database.}
\usage{
getPostgresqlConnection(
  user = getOption("GeneInfo2db.DB_USER", default = Sys.getenv("USER")),
  name = getOption("GeneInfo2db.DB_NAME"),
  host = getOption("GeneInfo2db.DB_HOST"),
  password = getOption("GeneInfo2db.DB_PASSWORD"),
  port = as.integer(getOption("GeneInfo2db.DB_PORT", default = 5432L))
)
}
\arguments{
\item{user}{database user}

\item{name}{database name}

\item{host}{database host}

\item{password}{database password}

\item{port}{database port (default 5432)}
}
\value{
connection to postgres database.
}
\description{
Get connection to postresql database.
}
\examples{

con <- getPostgresqlConnection()
DBI::dbGetQuery(con, "SELECT 1")
RPostgres::dbDisconnect(con)

}
