% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessedCopyNumber.R
\name{getTissueCopyNumber}
\alias{getTissueCopyNumber}
\title{Get Tissue Copy Number Variation Data}
\usage{
getTissueCopyNumber()
}
\value{
A list containing the data frame with processed copy number variation data
for specific genes.
}
\description{
Get Tissue Copy Number Variation Data
}
\examples{
result <- getTissueCopyNumber()
}
