% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueGSVA.R
\name{getTissueGSVA}
\alias{getTissueGSVA}
\title{Get Tissue Gene Set Variation Analysis (GSVA) Data}
\usage{
getTissueGSVA()
}
\value{
A list containing the data frame with GSVA scores for hallmark gene sets
across different tissues.
}
\description{
Get Tissue Gene Set Variation Analysis (GSVA) Data
}
