% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueMutation.R
\name{getTissueMutation}
\alias{getTissueMutation}
\title{Get Tissue Mutation Data}
\usage{
getTissueMutation()
}
\value{
A list containing the following element: a data frame with processed 
mutation data for specific transcripts and tissues.
}
\description{
This function connects to a PostgreSQL database and retrieves mutation data
for specific tissues.
}
