% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueProcessedRNASeq.R
\name{getTissueProcessedRNASeq}
\alias{getTissueProcessedRNASeq}
\title{Get Processed RNA-Seq Data for TCGA and GTEx Tissues}
\usage{
getTissueProcessedRNASeq(projects)
}
\description{
Get Processed RNA-Seq Data for TCGA and GTEx Tissues
}
