% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueProcessedRNASeq.R
\name{getTissueProcessedRNASeqProjects}
\alias{getTissueProcessedRNASeqProjects}
\title{Get Processed RNA-Seq Projects for TCGA and GTEx}
\usage{
getTissueProcessedRNASeqProjects()
}
\value{
A data frame with information on processed RNA-Seq projects for TCGA and GTEx.
}
\description{
This function retrieves a list of processed RNA-Seq projects for TCGA (The Cancer Genome Atlas)
and GTEx (Genotype-Tissue Expression) from the recount3 package's available projects data.
}
\examples{
result <- getTissueProcessedRNASeqProjects()

}
