% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessedCopyNumber.R
\name{getTissueProjectCopyNumber}
\alias{getTissueProjectCopyNumber}
\title{Get Tissue Project Copy Number Data}
\usage{
getTissueProjectCopyNumber(project, workflow = "ASCAT3")
}
\arguments{
\item{project}{A string. Name of the GDC project, e.g. "TCGA-ACC"}

\item{workflow}{A string. Name of GDC workflow to load data (default = "ASCAT3").
[Workflows Documentation](https://docs.gdc.cancer.gov/Data/Bioinformatics_Pipelines/CNV_Pipeline/)}
}
\value{
A data frame containing relative and absolute copy number information 
for genes in the specified tissue project.
}
\description{
Function to load copy number data from GDC
}
\examples{
result <- getTissueProjectCopyNumber("TCGA-ACC")

}
