% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueProcessedRNASeq.R
\name{getTissueRNAseqGroup}
\alias{getTissueRNAseqGroup}
\title{Get Tissue RNA-Seq Group Data}
\usage{
getTissueRNAseqGroup()
}
\value{
a data frame
}
\description{
This function generates a data frame representing RNA-Seq groups, including their IDs,
names, data file paths, and processing pipelines.
}
