% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueSignalingPathway.R
\name{getTissueSignalingPathway}
\alias{getTissueSignalingPathway}
\title{Get Tissue Signaling Pathway Data}
\usage{
getTissueSignalingPathway()
}
\value{
a list
}
\description{
This function retrieves signaling pathway data for specific tissues.
}
