% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeRead.R
\name{guessingReadingFunction}
\alias{guessingReadingFunction}
\title{Guess reading function from its content (e.g. csv, tab separated etc) and use that function to read a file.}
\usage{
guessingReadingFunction(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
a data frame resulting from read the file
}
\description{
Guess reading function from its content (e.g. csv, tab separated etc) and use that function to read a file.
}
