% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSettings.R
\name{pgpass2dbConfig}
\alias{pgpass2dbConfig}
\title{Get database config from .pgpass file}
\usage{
pgpass2dbConfig(line = 1, .pgpassPaths = c(".pgpass", "~/.pgpass"))
}
\arguments{
\item{line}{which line of .pgpass should be used}

\item{.pgpassPaths}{path to .pgpass file. Default value should work in most 
cases.}
}
\value{
list with database configuration.
}
\description{
Get database config from .pgpass file
}
\note{
To use specific config, it need to be set using \code{setDBconfig}.
}
