% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSettings.R
\name{print.GeneInfo2DatabaseConfig}
\alias{print.GeneInfo2DatabaseConfig}
\title{Method for printing GeneInfo2DatabaseConfig object.}
\usage{
\method{print}{GeneInfo2DatabaseConfig}(x, ...)
}
\arguments{
\item{x}{GeneInfo2DatabaseConfig object.}

\item{...}{not used.}
}
\description{
Method for printing GeneInfo2DatabaseConfig object.
}
\details{
this function hides the password value.
}
