% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileData.R
\name{registerGeneInfoDownloadFunction}
\alias{registerGeneInfoDownloadFunction}
\title{Registers GeneInfo2db default function used}
\usage{
registerGeneInfoDownloadFunction(
  fun = getFileDownload,
  version = getFileVersion()
)
}
\arguments{
\item{fun}{function used for downloading file.}
}
\value{
no value returned. Called for side effects.
}
\description{
Registers GeneInfo2db default function used
}
\details{
This function is for advanced users only. It allows to
change a function that is used to download source files.
}
