% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeRead.R
\name{safeReadFile}
\alias{safeReadFile}
\title{Safe read file from URL.}
\usage{
safeReadFile(
  URL,
  filename = NULL,
  read_fnc = readr::read_tsv,
  .retries = 20,
  .waitTime = 20,
  ...
)
}
\arguments{
\item{URL}{URL to file.}

\item{read_fnc}{function used to read file from disk. Default \code{\link{read_tsv}} from \code{readr} package.}

\item{...}{other parameres passed to \code{read_fnc}.}
}
\value{
A data.frame or tibble resulting from using the \code{read_fnc}.
}
\description{
Safe read file from URL.
}
\details{
This function is a safer alternative to \code{read_tsv(URL)}. When there will be any problem with the connection, \code{safe_read_file}
will fail, but the \code{read_tsv(URL)} can return partial result (meaning - all rows downloaded up to the loss connection point).
}
\examples{

\dontrun{
safeReadFile("https://ftp.ncbi.nlm.nih.gov/refseq/H_sapiens/Homo_sapiens.gene_info.gz")
}

}
