% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSettings.R
\name{setDBconfig}
\alias{setDBconfig}
\title{Set database config.}
\usage{
setDBconfig(host, port, name, user, pass)
}
\arguments{
\item{host}{database host or list with configuration (preferably the result
of pgpass2dbConfig)}

\item{port}{database port}

\item{name}{database name}

\item{user}{database user}

\item{pass}{database password}
}
\value{
nothing. Called for side effects.
}
\description{
Set database config.
}
\note{
the order of arguments mimics the order values in .pgpass file.
}
