% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFileCache.R
\name{setLocalFileRepo}
\alias{setLocalFileRepo}
\title{Set path to local file cache.}
\usage{
setLocalFileRepo(path)
}
\arguments{
\item{path}{path to local file cache.}
}
\description{
Set path to local file cache.
}
