% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFileCache.R
\name{useLocalFileRepo}
\alias{useLocalFileRepo}
\title{Modify file path to include local file cache.}
\usage{
useLocalFileRepo(path)
}
\arguments{
\item{path}{path to file to download.}
}
\value{
a file path consisting the local file chace path.

\dontrun{
useLocalFileRepo("xd.rds")
}
}
\description{
Modify file path to include local file cache.
}
