% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDatabase.R
\name{writeDatabase}
\alias{writeDatabase}
\title{writes a list of data frames into a Postgresql database}
\usage{
writeDatabase(table_data)
}
\arguments{
\item{table_data}{a list of data frames}
}
\value{
nothing
}
\description{
writes a list of data frames into a Postgresql database
}
